int 값이었던 점프값 float로 수정

+++ Assets/Scripts/AnimalDatabase/AnimalStatData.cs
@@ -9,9 +9,9 @@ public class AnimalStatData : ScriptableObject
     public string StringID;
     public int Grade;
     public int AttackPower;
-    public int StartSpeed;
-    public int MaxSpeed;
-    public int Jump;
+    public float StartSpeed;
+    public float MaxSpeed;
+    public float Jump;
 
     public void SetData(AnimalDataTable.AnimalData rawData)
     {
